Split emit_design_main into stage-owned assembly files
Extract realization descriptor emission from emit_design_main.cpp into
stage-owned files with a narrow public boundary. The public header
exports one orchestration function (EmitRealizationAndConstructor) and
one result type (RealizationEmissionResult) with semantically named
nested groups (ProcessMeta, DispatchMeta, ObservationMeta) reflecting
the runtime handoff contract.

File decomposition by responsibility:
- emit_realization_descriptors: constructor emission + orchestration
- emit_body_descriptors: per-body descriptor globals
- emit_connection_descriptors: connection descriptor globals
- emit_descriptor_utils: shared template/validation utilities
- emit_design_main: runtime entry/main emission only

All internal *Emission carrier structs and helper declarations are
confined to a private internal header. Schema, slot, and blob emission
utilities that are exclusively consumed by the orchestration file live
in its anonymous namespace rather than in the shared internal surface.
